Unreviewed. Race condition when closing pages with network process enabled https://bugs.webkit.org/show_bug.cgi?id=129684
Source/WebKit2:

When the last page of a web process is closed with the network
process enabled, the UI process checks if the web process can be
terminated and closes the connection with that web process. But
right before that the Close message has been sent which also
enables the process termination in the web process. Sometimes
there's a race condition in the web process between the process
termination triggered by the Close message and the one triggered
by the socket being closed by the UI process.

* WebProcess/WebProcess.cpp:
(WebKit::WebProcess::shouldTerminate): Use a 1 second timeout so
that if the message is sent right the before the connection is
closed, we don't wait indefinitely for the sync reply keeping the
web process alive.

Tools:

Enable the test to check that web processes finish when the web
view is destroyed.

* TestWebKitAPI/Tests/WebKit2Gtk/TestMultiprocess.cpp:

git-svn-id: http://svn.webkit.org/repository/webkit/releases/WebKitGTK/webkit-2.4@166161 268f45cc-cd09-0410-ab3c-d52691b4dbfc