Merge r176458 - Crash when setting 'transition-delay' CSS property to a calculated value https://bugs.webkit.org/show_bug.cgi?id=138784
Reviewed by Sam Weinig.

Source/WebCore:

Update CSSPrimitiveValue::computeTime() to use primitiveType() instead
of m_primitiveUnitType so that it properly handles calculated values.
Without this, we would hit the ASSERT_NOT_REACHED() assertion in
computeTime() for calculated values.

Test: fast/css/transition-delay-calculated-value.html

* css/CSSPrimitiveValue.h:
(WebCore::CSSPrimitiveValue::computeTime):

LayoutTests:

Add a layout test to check that setting the 'transition-delay' CSS
property to a calculated value does not crash and works as intended.

* fast/css/transition-delay-calculated-value-expected.txt: Added.
* fast/css/transition-delay-calculated-value.html: Added.

git-svn-id: http://svn.webkit.org/repository/webkit/releases/WebKitGTK/webkit-2.6@178270 268f45cc-cd09-0410-ab3c-d52691b4dbfc