<rdar://problem/17889529> CrashTracer: [USER] com.apple.WebKit.WebContent at com.apple.WebCore: WebCore::SharedBuffer::createPurgeableBuffer const + 14
Reviewed by NOBODY (OOPS!).

Speculative fix for the above crash. The current theory is that a timer is being fired for a
CachedResource that has been freed. Add a marker field in CachedResource to do an early
return from CachedResource::makePurgeable() when this happens.

No new tests because I can't reproduce the crash.

* loader/cache/CachedResource.cpp:
(WebCore::CachedResource::CachedResource):
(WebCore::CachedResource::~CachedResource):
(WebCore::CachedResource::makePurgeable):
* loader/cache/CachedResource.h:

git-svn-id: http://svn.webkit.org/repository/webkit/branches/safari-600.5-branch@179995 268f45cc-cd09-0410-ab3c-d52691b4dbfc