Merge r181392 - [GTK] Do not look for child processes in the UI process binary path https://bugs.webkit.org/show_bug.cgi?id=135752
Reviewed by Gustavo Noronha Silva.

.:

* Source/cmake/OptionsGTK.cmake: Add -DDEVELOPMENT_BUILD=1 to the
build for development builds.

Source/WebKit2:

It's only useful for internal tools and tests, but never when
installed, since we don't install the processes in the bin dir but
in the libexec dir.

* Shared/gtk/ProcessExecutablePathGtk.cpp:
(WebKit::findWebKitProcess): Only look or the executables in the
UI process binary path or WEBKIT_EXEC_PATH for development builds.
* UIProcess/API/gtk/WebKitWebContext.cpp:
(injectedBundleDirectory): Only check
WEBKIT_INJECTED_BUNDLE_PATH env var for development builds.

git-svn-id: http://svn.webkit.org/repository/webkit/releases/WebKitGTK/webkit-2.8@181529 268f45cc-cd09-0410-ab3c-d52691b4dbfc