Merge r188720 - Overflow check elimination fails for a simple test case https://bugs.webkit.org/show_bug.cgi?id=147387
Reviewed by Benjamin Poulain.

Source/JavaScriptCore:

Overflow check elimination was having issues when things got constant-folded, because whereas an
Add or LessThan operation teaches us about relationships between the things being added or
compared, we don't do that when we see a JSConstant. We don't create a relationship between every
JSConstant and every other JSConstant. So, if we constant-fold an Add, we forget the relationships
that it would have had with its inputs.

One solution would be to have every JSConstant create a relationship with every other JSConstant.
This is dangerous, since it would create O(n^2) explosion of relationships.

Instead, this patch teaches filtration and merging how to behave "as if" there were inter-constant
relationships. Normally those operations only work on two relationships involving the same node
pair. But now, if we have @x op @c and @x op @d, where @c and @d are different nodes but both are
constants, we will do merging or filtering by grokking the constant values.

This speeds up lots of tests in JSRegress, because it enables overflow check elimination on things
like:

for (var i = 0; i < 100; ++i)

Previously, the fact that this was all constants would throw off the analysis because the analysis
wouldn't "know" that 0 < 100.

* dfg/DFGIntegerRangeOptimizationPhase.cpp:

LayoutTests:

Added two test cases that previously would have an unnecessary overflow check on an induction
variable. These tests speed up by 10-15% thanks to this change.

Also added .html/expected files for some regress test that didn't have them.

* js/regress/function-call-expected.txt: Added.
* js/regress/function-call.html: Added.
* js/regress/hard-overflow-check-equal-expected.txt: Added.
* js/regress/hard-overflow-check-equal.html: Added.
* js/regress/hard-overflow-check-expected.txt: Added.
* js/regress/hard-overflow-check.html: Added.
* js/regress/script-tests/hard-overflow-check-equal.js: Added.
(foo):
* js/regress/script-tests/hard-overflow-check.js: Added.
(foo):

git-svn-id: http://svn.webkit.org/repository/webkit/releases/WebKitGTK/webkit-2.10@189691 268f45cc-cd09-0410-ab3c-d52691b4dbfc