Merge r191541 - [TexMap] Clean up BitmapTexture and BitmapTextureGL. https://bugs.webkit.org/show_bug.cgi?id=143298
Reviewed by Žan Doberšek.

No new tests, this is just a refactor.

* platform/graphics/texmap/BitmapTexture.h:
(WebCore::BitmapTexture::canReuseWith): Deleted.
Reuseability of a BitmapTexture is only decided by its size.
We can use size() instead of this method.

* platform/graphics/texmap/BitmapTextureGL.h:
(WebCore::driverSupportsExternalTextureBGRA): Deleted.
We do not have to check a suitable texture format in every
update/reset. It is enough to store the formats in construction time
and reuse them.

git-svn-id: http://svn.webkit.org/repository/webkit/releases/WebKitGTK/webkit-2.10@191579 268f45cc-cd09-0410-ab3c-d52691b4dbfc