Merge r196061 - [GTK] Layout Test http/tests/appcache/different-https-origin-resource-main.html is failing https://bugs.webkit.org/show_bug.cgi?id=145253
Reviewed by Michael Catanzaro.

Source/WebCore:

The problem is that when the load is cancelled while the
connection is still being established,
SoupMessage::notify::tls-errors is emitted and the handler calls
ResourceHandleClient::didFail() which can delete the ResourceHandle.

* platform/network/soup/ResourceHandleSoup.cpp:
(WebCore::tlsErrorsChangedCallback): Protect the ResourceHandle
for the scope of the callback because
ResourceHandleClient::didFail() could delete the object.

LayoutTests:

Update test expectation.

* platform/gtk/TestExpectations:

git-svn-id: http://svn.webkit.org/repository/webkit/releases/WebKitGTK/webkit-2.10@196911 268f45cc-cd09-0410-ab3c-d52691b4dbfc