Merge r196631 - Mouse cursor doesn't change when entering scrollbars https://bugs.webkit.org/show_bug.cgi?id=154243
Reviewed by Simon Fraser.

If the scrollbar is over or very close to text or a link, when
entering the scrollbar the cursor is not changed, keeping the beam
or hand cursor when using the scrollbar. Same happens for image
documents where the magnifier cursor is used and it remains when
entering the scrollbars. We should use pointer cursor always for
scrollbars.

* page/EventHandler.cpp:
(WebCore::EventHandler::updateCursor): Request also to include
frame scrollbars in hit test result.
(WebCore::EventHandler::selectCursor): Use always pointer cursor
for scrollbars.

git-svn-id: http://svn.webkit.org/repository/webkit/releases/WebKitGTK/webkit-2.10@196928 268f45cc-cd09-0410-ab3c-d52691b4dbfc