Merge r198498 - Very flashy scrolling on http://quellish.tumblr.com page https://bugs.webkit.org/show_bug.cgi?id=155728 rdar://problem/22299375
Reviewed by Zalan Bujtas.

http://quellish.tumblr.com/post/126712999812/how-on-earth-the-facebook-ios-application-is-so
has many elements that are nested inside elements with non-equal corner radius clipping.
This requires building bezier paths for the rounded-rect clip which is expensive.

For many rows of the table, we can avoid the rounded-rect clipping because the intersection
of the paintDirtyRect and the clip is actually rectangular.

* platform/graphics/FloatRoundedRect.cpp:
(WebCore::FloatRoundedRect::intersectionIsRectangular):
* platform/graphics/FloatRoundedRect.h:
* rendering/RenderLayer.cpp:
(WebCore::RenderLayer::clipToRect):

git-svn-id: http://svn.webkit.org/repository/webkit/releases/WebKitGTK/webkit-2.12@199426 268f45cc-cd09-0410-ab3c-d52691b4dbfc