Merge r204936 - Assertion hit: ASSERT(!m_parserScheduler) in ~HTMLDocumentParser() https://bugs.webkit.org/show_bug.cgi?id=161083 <rdar://problem/27986937>
Reviewed by Darin Adler.

Source/WebCore:

Call cancelParsing() right before re-creating the new parser in
Document::implicitOpen(). We would previously call removeChildren()
in between which is was an issue because it can cause JS to be
executed and therefore the parser to be re-created after we have
cancelled parsing.

No new tests, unskipped html/browsers/browsing-the-web/unloading-documents/004.html
which was consistently crashing.

* dom/Document.cpp:
(WebCore::Document::implicitOpen):

LayoutTests:

Unskip html/browsers/browsing-the-web/unloading-documents/004.html as
it is no longer crashing.

* TestExpectations:

git-svn-id: http://svn.webkit.org/repository/webkit/releases/WebKitGTK/webkit-2.12@205435 268f45cc-cd09-0410-ab3c-d52691b4dbfc