Merge r205970 - Cleanup RenderLayer::shouldBeNormalFlowOnly https://bugs.webkit.org/show_bug.cgi?id=161981
Reviewed by Simon Fraser.

Source/WebCore:

This patch changes the logic of figuring about if a particular layer is normal flow only by simply checking
if the layer creates a stacking context. If it does, we assume that it can't be a normal flow layer anymore.
This patch slightly changes behaviour by making layers with isolation and reflection to be non normal flow layers anymore.

Covered by existing testcases.

* rendering/RenderLayer.cpp:
(WebCore::RenderLayer::calculateClipRects):

LayoutTests:

Rebaselining.

* fast/table/overflow-table-collapsed-borders-section-self-painting-layer-painting-expected.txt:

git-svn-id: http://svn.webkit.org/repository/webkit/releases/WebKitGTK/webkit-2.14@207076 268f45cc-cd09-0410-ab3c-d52691b4dbfc