Merge r206513 - [GTK] Simplify platformForUAString https://bugs.webkit.org/show_bug.cgi?id=162614
Reviewed by Carlos Garcia Campos.

Note in particular that the previous PLATFORM(MAC) check was wrong here, as that is never
true for GTK; this patch risks changing it to OS(MAC_OS_X), on the hope that advertising Mac
when running on Mac might not break anything. If it does, then we should remove it and just
always pretend to be on X11.

* platform/gtk/UserAgentGtk.cpp:
(WebCore::platformForUAString):

git-svn-id: http://svn.webkit.org/repository/webkit/releases/WebKitGTK/webkit-2.14@207117 268f45cc-cd09-0410-ab3c-d52691b4dbfc