Merge r213990 - [GTK] Initialize m_button, m_clickCount members in PlatformMouseEvent constructors https://bugs.webkit.org/show_bug.cgi?id=169666
Reviewed by Michael Catanzaro.

Initialize the m_button and m_clickCount class members in the GTK+-specific
implementation of PlatformMouseEvent constructors to NoButton and 0,
respectively. The constructors expect to operate on passed-in GTK+ events
that will be able to initialize those two members to some valid values, but
this is not guaranteed.

* platform/gtk/PlatformMouseEventGtk.cpp:
(WebCore::PlatformMouseEvent::PlatformMouseEvent):

git-svn-id: http://svn.webkit.org/repository/webkit/releases/WebKitGTK/webkit-2.16@214742 268f45cc-cd09-0410-ab3c-d52691b4dbfc