Merge r222544 - AX: Several ARIA roles with presentational children are exposing children https://bugs.webkit.org/show_bug.cgi?id=177502
Reviewed by Chris Fleizach.

Source/WebCore:

Add MenuItemCheckboxRole, MenuItemRadioRole, and SplitterRole to the list
of AccessibilityRole types which cannot have children. In the case of
DocumentMathRole, only treat children as presentational (excluded from the
accessibility tree) if the element is not MathML.

Test: accessibility/presentational-children.html

* accessibility/AccessibilityNodeObject.cpp:
(WebCore::AccessibilityNodeObject::canHaveChildren const):

LayoutTests:

New test. Also update the results for the inspector because ARIA separators
and their subclass roles should not expose (claim to have) children.

* accessibility/presentational-children-expected.txt: Added.
* accessibility/presentational-children.html: Added.
* inspector/dom/getAccessibilityPropertiesForNode-expected.txt: Updated.

git-svn-id: http://svn.webkit.org/repository/webkit/releases/WebKitGTK/webkit-2.18@223467 268f45cc-cd09-0410-ab3c-d52691b4dbfc