Merge r222604 - [Cairo] Remove unnecessary cairo.h includes https://bugs.webkit.org/show_bug.cgi?id=177540
Reviewed by Carlos Garcia Campos.

Drop unnecessary includes of the cairo.h header from the Font.h,
FontPlatformData.h and Pattern.h headers. For the former two the
header isn't required, while in the last one we can use a forward
declaration for the cairo_pattern_t type instead of including
the whole Cairo headers to have that available.

* platform/graphics/Font.h:
* platform/graphics/FontPlatformData.h:
* platform/graphics/Pattern.h:

git-svn-id: http://svn.webkit.org/repository/webkit/releases/WebKitGTK/webkit-2.18@223479 268f45cc-cd09-0410-ab3c-d52691b4dbfc