AX: [ATK] Explicitly-set aria-sort value of "none" should be exposed as an object attribute https://bugs.webkit.org/show_bug.cgi?id=177955
Reviewed by Chris Fleizach.

Source/WebCore:

Expose "none" as the value of aria-sort when the attribute is present and not
one of the other valid sort types. Also add a new AccessibilitySortDirection,
SortDirectionInvalid. This is currently only being used when the sort direction
is sought for a role which does not support this ARIA attribute, but might come
in handy if we want to add further sanity checks on the author-provided values.

Test: accessibility/gtk/aria-sort-values.html

* accessibility/AccessibilityObject.cpp:
(WebCore::AccessibilityObject::sortDirection const):
* accessibility/AccessibilityObject.h:
* accessibility/atk/WebKitAccessibleWrapperAtk.cpp:
(webkitAccessibleGetAttributes):
* accessibility/mac/WebAccessibilityObjectWrapperMac.mm:
(-[WebAccessibilityObjectWrapper additionalAccessibilityAttributeNames]):

LayoutTests:

* accessibility/gtk/aria-sort-values-expected.txt: Added.
* accessibility/gtk/aria-sort-values.html: Added.

git-svn-id: http://svn.webkit.org/repository/webkit/releases/WebKitGTK/webkit-2.18@223522 268f45cc-cd09-0410-ab3c-d52691b4dbfc