Merge r228952 - ArgumentsEliminationPhase has a branch on GetByOffset that should be an assert https://bugs.webkit.org/show_bug.cgi?id=182982
Reviewed by Yusuke Suzuki.

I don't know why this check was not always an assert. When we see
a GetByOffset on an eliminated allocation, that allocation *must*
be a PhantomClonedArguments. If it weren't, the GetByOffset would
have escaped it. Because this transformation happens by visiting
blocks in pre-order, and by visiting nodes in a block starting from
index zero to index block->size() - 1, we're guaranteed that eliminated
allocations get transformed before users of it, since we visit nodes
in dominator order.

* dfg/DFGArgumentsEliminationPhase.cpp:

git-svn-id: http://svn.webkit.org/repository/webkit/releases/WebKitGTK/webkit-2.20@229234 268f45cc-cd09-0410-ab3c-d52691b4dbfc