Merge r230264 - JSArray::appendMemcpy seems to be missing a barrier https://bugs.webkit.org/show_bug.cgi?id=184290
Reviewed by Mark Lam.

If you write to an array that may contain pointers and you didn't just allocate it, then you need to
barrier right after.

I don't know if this is really a bug - it's possible that all callers of appendMemcpy do things that
obviate the need for this barrier. But these barriers are cheap, so we should do them if in doubt.

* runtime/JSArray.cpp:
(JSC::JSArray::appendMemcpy):

git-svn-id: http://svn.webkit.org/repository/webkit/releases/WebKitGTK/webkit-2.20@230432 268f45cc-cd09-0410-ab3c-d52691b4dbfc