Merged r237274 - [MSE] timestampOffset can introduce floating-point rounding errors to incoming samples https://bugs.webkit.org/show_bug.cgi?id=190590 <rdar://problem/45275626>
Reviewed by Eric Carlson.

Source/WebCore:

Test: media/media-source/media-source-timestampoffset-rounding-error.html

SourceBuffer.timestampOffset is a double property, which, when added to a MediaTime will
result in a double-backed MediaTime as PTS & DTS. This can introduce rounding errors when
these samples are appended as overlapping existing samples. Rather than converting a MediaTime
to double-backed when adding the timestampOffset, convert the offset to a multiple of the
sample's timeBase.

* Modules/mediasource/SourceBuffer.cpp:
(WebCore::SourceBuffer::setTimestampOffset):
(WebCore::SourceBuffer::sourceBufferPrivateDidReceiveSample):

LayoutTests:

* media/media-source/media-source-sequence-timestamps-expected.txt:
* media/media-source/media-source-timestampoffset-rounding-error-expected.txt: Added.
* media/media-source/media-source-timestampoffset-rounding-error.html: Added.
* media/media-source/mock-media-source.js:
(makeASample):

git-svn-id: http://svn.webkit.org/repository/webkit/releases/WebKitGTK/webkit-2.22@237514 268f45cc-cd09-0410-ab3c-d52691b4dbfc