Merge r236969 - StringTypeAdapter constructor is not properly enforcing String::MaxLength. https://bugs.webkit.org/show_bug.cgi?id=190392 <rdar://problem/45116210>
Reviewed by Saam Barati.

Previously, the StringTypeAdapter constructor for a UChar* string was summing the
unsigned length of the source string without an overflow check.  We now make that
length a size_t which removes this issue, and assert that it's within
String::MaxLength thereafter.

Also made the StringTypeAdapter constructor for a LChar* string behave in an
equivalent manner for consistency.  In both cases, we'll crash in a RELEASE_ASSERT
if the source string length exceeds String::MaxLength.

* wtf/text/StringConcatenate.h:

git-svn-id: http://svn.webkit.org/repository/webkit/releases/WebKitGTK/webkit-2.22@238974 268f45cc-cd09-0410-ab3c-d52691b4dbfc