Merged r244108 - [WPE] Avoid async IO starving timers https://bugs.webkit.org/show_bug.cgi?id=196733
Reviewed by Carlos Garcia Campos.

If AsyncIONetwork and DiskCacheRead priorities are higher than
MainThreadSharedTimer the timers get starved. This causes the NetworkProcess
to accumulate MB of data instead of handing it down to the WebProcess (done
using a Timer). This eventually causes an Out Of Memory kill on the
NetworkProcess on some embedded platforms with low memory limits.

This patch levels the three priorities to the same value, while still leaving
DiskCacheWrite with less priority than DiskCacheRead.

* wtf/glib/RunLoopSourcePriority.h: Changed RunLoopSourcePriority values for WPE.

git-svn-id: http://svn.webkit.org/repository/webkit/releases/WebKitGTK/webkit-2.24@244452 268f45cc-cd09-0410-ab3c-d52691b4dbfc