Unreviewed build fix for missing WebCore library when linking TestNetscapePlugIn
Using LDFLAGS='-Wl,--no-undefined' would cause TestNetscapePlugIn to
fail linking as it uses WTF::Sleep() but the WTF library is not
directly listed for linking. While the library would be linked
indirectly, passing --no-undefined to the linker disallows that, so
it is needed to list it explicitly for it to succeed.

* DumpRenderTree/TestNetscapePlugIn/CMakeLists.txt: Add dependency on
the WTF library.

git-svn-id: http://svn.webkit.org/repository/webkit/releases/WebKitGTK/webkit-2.24@248870 268f45cc-cd09-0410-ab3c-d52691b4dbfc